+++ packages/plugins/Attributions/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Feature: Add link to repository and license as required by EUPL v1.2.
 - Feature: Include title in AttributionContent so users receive this information on all devices.
-- Fix: Attributions are now being translated when the window is open
+- Feature: `max-width` and `width` are now only set on the content if `renderType` is `independent` as the mobile view is rendered in the `MoveHandle`.
+- Fix: Attributions are now being translated when the window is open.
 
 ## 1.0.0